Fix SystemUI animator leak
When there was a state change or layout when there was an animation
set as the mobile icon in the status bar, we never stopped the old
animation, which was infinite. This was using more and more CPU because
the animations never got stopped.

To fix this, we don't update the drawable when a layout happens and we
stop the previous animation when replacing the icon.

Bug: 26616870
Change-Id: If501155d1a99d587e50a1b77ebb03a21c940662b

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/SignalClusterView.java